requirements.txt
pillow
requests


